stop-place-tile * Move ToggleDetailsPanel into StopPlaceTile directory
