+++ src/containers/Admin/EditTab/StopPlaceTile/StopPlaceTile.tsx
@@ -1,9 +1,9 @@
 import React, { useCallback } from 'react'
 import { xor } from 'lodash'
 import { Heading3 } from '@entur/typography'
-import { ToggleDetailsPanel } from '../ToggleDetailsPanel/ToggleDetailsPanel'
 import { EditTile } from '../EditTile/EditTile'
 import { useSettings } from '../../../../settings/SettingsProvider'
+import { ToggleDetailsPanel } from './ToggleDetailsPanel/ToggleDetailsPanel'
 import { StopPlaceSearch } from './StopPlaceSearch/StopPlaceSearch'
 import { StopPlacePanel } from './StopPlacePanel/StopPlacePanel'
 import classes from './StopPlaceTile.module.scss'
