+++ src/containers/Admin/EditTab/StopPlaceTile/ToggleDetailsPanel/ToggleDetailsPanel.scss
